refactor: SSE 레이어 의존성 분리 및 네이밍 개선
- SseEmitterService 도입으로 Controller → Store 직접 의존 제거
- heartbeat 스케줄링을 Repository에서 Service로 이동
- SseEmitterRepository → SseEmitterStore 리네이밍
